Started the Database implementation
I started with the database.  SInce we have not begun actually using any
of the data, and still working on the UI, I just made create tables
statements to be on create method, and to close the database on destroy.
I also wrote code for adding in data to the tables.  I was not able to
make any queries yet.  I also changed Income and expense to be a
subclasses of Transaction and some more methods in each.  I didn't test
the code bc I didn't get the chance, so if it crashes the app just go
ahead and comment it out.  SQLite uses different data fields than SQL so
I may have to remodel some of the data attributes to make it work.